+++ src/pages/logIn/form/LogInForm.tsx
@@ -1,28 +1,14 @@
-import {useLoginForm,type LoginFormValues}  from '@pages/logIn/form/useLoginForm';
+import {useLogInForm}  from '@pages/logIn/hook/useLogInForm';
 import InputBar from '@shared/components/inputBar/InputBar';
 import Text from '@shared/components/text/Text';
 import * as styles from '@pages/logIn/LogIn.css';
 import { useState } from 'react';
 import Button from '@shared/components/button/Button';
 
-const LoginForm = () => {
-  const { register, handleSubmit, formState: { errors }, setFocus, watch } = useLoginForm();
+export default function LoginForm() {
+  const { register, handleSubmit, onSubmit, onSubmitError, emailValue, passwordValue } = useLogInForm();
   const [checked, setChecked] = useState(false);
-  const emailValue = watch('email');
-  const passwordValue = watch('password');
 
-  const onSubmit = (data: LoginFormValues) => {
-    console.log('Form submitted with data:', data);
-  };
-  const onSubmitError = () => {
-    if (errors.email) {
-      setFocus('email');
-      console.log(errors.email.message);
-    } else if (errors.password) {
-      setFocus('password');
-      console.log(errors.password.message);
-    }
-  };
 
   return (
     <form onSubmit={handleSubmit(onSubmit, onSubmitError)} className={styles.inputContainer}>
@@ -52,7 +38,7 @@ const LoginForm = () => {
         <Text tag="body_14" color="blue1">자동 로그인 설정</Text>
       </div>
       <div className={styles.buttonContainer}>
-        <Button text="로그인" size='long' variant='enabled' onClick={handleSubmit(onSubmit, onSubmitError)} />
+        <Button text="로그인" size='long' variant='enabled' onClick={handleSubmit(onSubmit, onSubmitError)}/>
         <div className={styles.signUpButton}>
           <div className={styles.line}/>
           <Text tag="body_14" color="blue1">회원가입</Text>
@@ -62,5 +48,3 @@ const LoginForm = () => {
     </form>
   );
 };
-
-export default LoginForm;
